update
Continued working on the AI implementation. Managed to make it run, I
have found a relevant forum message with our scenario. We can use the
Academy.Instance singleton to manipulate the MLAgents settings and
disable automatic stepping.

Currently the code runs but spawns infinite customers and hangs. I think
this might be because the Task that we await finished before the store
can Enqueue in parallel or something like that...

Also I have edited the previous scripts to be able to handle the new
Scene like CustomerController not setting a path, the way I check for
it is by validating that there is no CustomerManager. (this is very
specific but works for now. See code changes)